Build fix #4
